backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pymupdf>=1.23.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
openai>=1.10.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiosqlite>=0.19.0
httpx>=0.26.0
numpy>=1.24.0
aiofiles>=23.2.1
rank-bm25>=0.2.2
python-docx>=1.1.0
python-pptx>=0.6.23
slowapi>=0.1.9
redis>=5.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
easyocr>=1.7.0
docling>=2.0.0
